Move BatchTableHierarchyPropertyValues to a separate file.
